requirements.txt
black==24.3.0
heatshrink2==0.13.0
pillow==10.3.0
ffmpeg-python==0.2.0
